photos-to-s3/requirements.txt
google-api-python-client==1.6.7
python-dateutil==2.6.1
requests==2.18.4
